requirements.txt
# This file is autogenerated by pip-compile with `--generate-hashes`
# To update, run:
#
#    pip-compile requirements.in > requirements.txt
#    This is to be updated as we go and to be considered as a placeholder awaiting engineering's approval


# Core ML & Data Libraries
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2


# Text Embedding
sentence-transformers==2.2.2
torch==2.2.2
transformers==4.39.3


# Google Cloud SDKs
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-auth==2.29.0


# Validation & Utilities
pydantic==2.6.4
python-dotenv==1.0.1
rich==13.7.1


# Logging
loguru==0.7.2


# CLI tooling
click==8.1.7


# YAML config support
pyyaml==6.0.1


# Testing
pytest==8.1.1
pytest-mock==3.14.0